move login tests to test_api.py
