+++ tests/test_api.py
@@ -25,14 +25,68 @@ from dashlive.server import models
 from dashlive.server.options.repository import OptionsRepository
 from dashlive.server.options.types import CgiOption
 from dashlive.server.requesthandler.streams import ViewStreamAjaxResponse
+from dashlive.server.requesthandler.user_management import LoginResponseJson
 from dashlive.utils.date_time import to_iso_datetime
 from dashlive.utils.json_object import JsonObject
 
 from .mixins.flask_base import FlaskTestBase
+from .mixins.mock_time import MockTime
 from .mixins.stream_fixtures import BBB_FIXTURE
 
 class TestRestApi(FlaskTestBase):
-    def test_add_stream(self):
+    def test_login_unknown_user(self) -> None:
+        self.check_login_failure('not.known', 'password')
+
+    def test_login_wrong_password(self) -> None:
+        self.check_login_failure(self.STD_USER, 'password')
+        self.check_login_failure(self.STD_USER, self.ADMIN_PASSWORD)
+
+    def check_login_failure(self, username: str, password: str) -> None:
+        url = flask.url_for('api-login')
+        data = {
+            'username': username,
+            'password': password,
+        }
+        response = self.client.post(url, json=data, headers={
+            'content-type': 'application/json',
+        })
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual({
+            "success": False,
+            "error": "Wrong username or password",
+            "csrf_token": response.json['csrf_token'],
+        }, response.json)
+
+    @MockTime("2023-07-18T20:10:02Z")
+    def test_ajax_login_page(self) -> None:
+        url = flask.url_for('api-login')
+        payload = {
+            'username': self.STD_USER,
+            'password': self.STD_PASSWORD,
+            'rememberme': False,
+        }
+        response = self.client.post(url, json=payload)
+        self.assert200(response)
+        js_response: LoginResponseJson = response.json
+        self.assertNotIn('error', js_response)
+        expected = {
+            'csrf_token': js_response['csrf_token'],
+            'success': True,
+            'accessToken': js_response['accessToken'],
+            'refreshToken': js_response['refreshToken'],
+            'mustChange': False,
+            'user': {
+                'email': self.STD_EMAIL,
+                'groups': ['USER'],
+                'last_login': '2023-07-18T20:10:02Z',
+                'pk': 2,
+                'username': self.STD_USER,
+            }
+        }
+        self.maxDiff = None
+        self.assertDictEqual(expected, js_response)
+
+    def test_add_stream(self) -> None:
         self.assertEqual(models.Stream.count(), 0)
         request = {
             'title': 'Big Buck Bunny',
